requirements.txt
# AC Charger Project Requirements
# Created: 2025-12-08
# Last Modified: 2025-12-09
# Version: 1.1.0

# ESP32 Communication
pyserial>=3.5

# Modbus (Meter)
pymodbus==3.6.7

# REST API Framework
fastapi>=0.124.0
uvicorn[standard]>=0.38.0
pydantic>=2.12.5
slowapi>=0.1.9

# Testing (development)
pytest>=9.0.0
pytest-asyncio>=1.3.0
pytest-mock>=3.15.0
pytest-cov>=7.0.0
coverage>=7.0.0
hypothesis>=6.0.0
pytest-benchmark>=5.0.0

# Code Quality (development)
# black>=24.0.0
# ruff>=0.5.0
# mypy>=1.11.0

# Environment Variables
python-dotenv>=1.0.0

# System Monitoring
psutil>=5.9.0

# Prometheus Metrics
prometheus-client>=0.21.0

# Code Quality Tools (development)
black>=24.0.0
ruff>=0.14.0

# Caching (optional - Redis için)
# redis>=5.0.0  # Uncomment if using Redis backend

